requirements.txt
click==7.1.1
requests==2.23.0
PyYAML==5.3.1
devo-sdk==3.3.2
Jinja2==2.11.1
psutil==5.7.0
python-dateutil==2.8.1
Faker==4.0.2
pycron==3.0.0
